+++ extensions/notepadia/src/browser/notepadia-frontend-module.ts
@@ -12,6 +12,7 @@ import { NotepadiaContribution } from './notepadia-contribution';
 import { NotepadiaDropContribution } from './notepadia-drop-contribution';
 import { NotepadiaEditorKeybindingContribution } from './notepadia-editor-keybinding-contribution';
 import { NotepadiaEncodingContribution } from './notepadia-encoding-contribution';
+import { NotepadiaEolContribution } from './notepadia-eol-contribution';
 import { NotepadiaKeybindingContribution } from './notepadia-keybinding-contribution';
 import { NotepadiaMenuContribution } from './notepadia-menu-contribution';
 import { NotepadiaStatusBarContribution } from './notepadia-status-bar-contribution';
@@ -23,6 +24,10 @@ export default new ContainerModule((bind) => {
     bind(NotepadiaEncodingContribution).toSelf().inSingletonScope();
     bind(CommandContribution).toService(NotepadiaEncodingContribution);
     bind(MenuContribution).toService(NotepadiaEncodingContribution);
+
+    bind(NotepadiaEolContribution).toSelf().inSingletonScope();
+    bind(CommandContribution).toService(NotepadiaEolContribution);
+    bind(MenuContribution).toService(NotepadiaEolContribution);
 
     bind(NotepadiaDropContribution).toSelf().inSingletonScope();
     bind(FrontendApplicationContribution).toService(NotepadiaDropContribution);